fix: osc client compat (countdowns)
Former-commit-id: 65e0bf9f4d157c82a88b7b7d1e2051e61394039d

+++ src/core/mixer/audio/audio_mixer.cpp
@@ -219,7 +219,7 @@ struct audio_mixer::impl : boost::noncopyable
         const int num_channels = format_desc_.audio_channels;
 
         state_.clear();
-        state_["nb_channels"] = num_channels;
+        state_["nb_channels"] = static_cast<int32_t>(num_channels);
 
         auto max = std::vector<int32_t>(num_channels, std::numeric_limits<int32_t>::min());
